api/requirements.txt
protobuf==4.21.1
python-dotenv==0.20.0
